Add equilibrium goals to bindings // fix MocoControlGoal exponent type

+++ Moco/Moco/MocoGoal/MocoControlGoal.h
@@ -69,7 +69,7 @@ public:
             const std::string& controlName, const double& weight);
 
     /// Set the exponent on the control signals.
-    void setExponent(double exponent) { set_exponent(exponent); }
+    void setExponent(int exponent) { set_exponent(exponent); }
     double getExponent() const { return get_exponent(); }
 
     /// Set if the goal should be divided by the displacement of the system's
